Lichess.data package restructure
Utility was becoming a bit of a dumping ground. I've moved all things
Lichess API related into lichess.data.
