+++ src/main/java/com/chessmates/lichess/data/LichessApiImpl.groovy
@@ -1,8 +1,9 @@
-package com.chessmates.utility
+package com.chessmates.lichess.data
 
 import com.chessmates.model.GameColor
 import com.chessmates.model.Game
 import com.chessmates.model.Player
+import com.chessmates.utility.HttpUtility
 import groovy.json.JsonSlurper
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
