+++ src/main/java/com/chessmates/lichess/data/LichessDataService.groovy
@@ -1,4 +1,4 @@
-package com.chessmates.service
+package com.chessmates.lichess.data
 
 import com.chessmates.model.Game
 import com.chessmates.model.Player
