+++ src/main/java/com/chessmates/service/EntityServiceImpl.groovy
@@ -1,5 +1,6 @@
 package com.chessmates.service
 
+import com.chessmates.lichess.data.LichessDataService
 import com.chessmates.model.Game
 import com.chessmates.model.Player
