+++ src/main/java/com/chessmates/utility/GetPageFunction.groovy
@@ -1,5 +1,7 @@
 package com.chessmates.utility
 
+import com.chessmates.lichess.data.LichessResultPage
+
 import java.util.function.Function
 
 /**
