+++ src/test/java/com/chessmates/lichess/data/LichessResultSetTest.groovy
@@ -1,5 +1,8 @@
-package com.chessmates.utility
+package com.chessmates.lichess.data
 
+import com.chessmates.lichess.data.LichessResultPage
+import com.chessmates.lichess.data.LichessResultSet
+import com.chessmates.utility.GetPageFunction
 import spock.lang.Specification
 
 import java.util.function.Predicate
